feat: auto generate message id if magic value "generate_id" is specified
If a magic `generate_id` value is specified in the message id in the wrapper, it gets replaced with an automatically generated value.

This was a feature available in ada_2_hl7/peri/2_2/kernset_aanleverbericht/wrapper/kernset_aanleverbericht_wrapper.xsl, but was not yet available in 2_3. This PR ports this functionality over.